# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CoinHub API - virtual coin economy platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkResponse {
  ok: boolean;
}

export interface User {
  id: string;
  publicId: string;
  username: string;
  coins: number;
  createdAt: string;
  isAdmin: boolean;
  email?: string | null;
  bio?: string | null;
  avatarColor: string;
  bonusReady: boolean;
  unreadNotifications: number;
  unreadDms: number;
}

export interface AuthResponse {
  user: User;
}

export interface Transaction {
  id: string;
  userId: string;
  amount: number;
  reason: string;
  source: string;
  createdAt: string;
}

export interface UserStats {
  totalEarned: number;
  totalSpent: number;
  gamesPlayed: number;
  rank?: number | null;
}

export interface WheelSegment {
  multiplier: number;
  label: string;
  color: string;
  rarity: string;
}

export interface GamesConfig {
  minBet: number;
  maxBet: number;
  wheelSegments: WheelSegment[];
  slotSymbols: string[];
}

export interface SlotResult {
  bet: number;
  /**
   * @minItems 3
   * @maxItems 3
   */
  symbols: string[];
  multiplier: number;
  won: number;
  netChange: number;
  newBalance: number;
  label: string;
  rarity: string;
  outcome: string;
}

export interface SpinResult {
  bet: number;
  segmentIndex: number;
  multiplier: number;
  won: number;
  netChange: number;
  newBalance: number;
  label: string;
  rarity: string;
}

export interface LuckyBoxReveal {
  multiplier: number;
  rarity: string;
}

export interface LuckyBoxResult {
  bet: number;
  pickIndex: number;
  /**
   * @minItems 9
   * @maxItems 9
   */
  boxes: LuckyBoxReveal[];
  multiplier: number;
  won: number;
  netChange: number;
  newBalance: number;
  label: string;
  rarity: string;
}

export interface CrashResult {
  bet: number;
  autoCashout: number;
  crashAt: number;
  multiplier: number;
  won: number;
  netChange: number;
  newBalance: number;
  label: string;
  rarity: string;
  cashedOut: boolean;
}

export interface ChatMessage {
  id: string;
  userId: string;
  publicId: string;
  username: string;
  avatarColor: string;
  isAdmin: boolean;
  message: string;
  createdAt: string;
}

export interface PublicUser {
  id: string;
  publicId: string;
  username: string;
  avatarColor: string;
  isAdmin: boolean;
}

export interface PublicProfile {
  id: string;
  publicId: string;
  username: string;
  avatarColor: string;
  isAdmin: boolean;
  bio?: string | null;
  coins: number;
  rank?: number | null;
  gamesPlayed: number;
  createdAt: string;
  /** none | pending_outgoing | pending_incoming | friends | self */
  friendStatus: string;
}

export interface TransferResult {
  ok: boolean;
  senderBalance: number;
  recipient: PublicUser;
  amount: number;
}

export interface BonusResult {
  granted: boolean;
  amount: number;
  newBalance: number;
  nextAvailableAt?: string | null;
}

export interface Notification {
  id: string;
  title: string;
  body: string;
  kind: string;
  readAt?: string | null;
  createdAt: string;
}

export interface FriendEntry {
  user: PublicUser;
  since: string;
}

export interface FriendList {
  friends: FriendEntry[];
  incoming: FriendEntry[];
  outgoing: FriendEntry[];
}

export interface DmThread {
  user: PublicUser;
  lastMessage: string;
  lastAt: string;
  unread: number;
}

export interface DmMessage {
  id: string;
  fromId: string;
  toId: string;
  message: string;
  createdAt: string;
  readAt?: string | null;
}

export interface NewsPost {
  id: string;
  title: string;
  body: string;
  createdAt: string;
}

export interface LeaderboardEntry {
  rank: number;
  publicId: string;
  username: string;
  coins: number;
}

export interface AdminStatus {
  isAdmin: boolean;
}

export interface AdminUser {
  id: string;
  publicId: string;
  username: string;
  coins: number;
  createdAt: string;
}

export interface AdminUserDetail {
  user: AdminUser;
  transactions: Transaction[];
}

export interface AdminStats {
  totalUsers: number;
  totalCoinsInCirculation: number;
  totalTransactions: number;
  coinsAddedToday: number;
}

export type RegisterUserBody = {
  /**
   * @minLength 3
   * @maxLength 24
   */
  username: string;
  /**
   * @minLength 4
   * @maxLength 64
   */
  password: string;
  email?: string;
};

export type LoginUserBody = {
  username: string;
  password: string;
};

export type UpdateMyProfileBody = {
  /** @maxLength 200 */
  bio?: string;
  avatarColor?: string;
  email?: string;
};

export type TransferCoinsBody = {
  recipientPublicId: string;
  /** @minimum 1 */
  amount: number;
  /** @maxLength 100 */
  note?: string;
};

export type SearchUsersParams = {
  q: string;
};

export type SendFriendRequestBody = {
  publicId: string;
};

export type SendDmMessageBody = {
  /**
   * @minLength 1
   * @maxLength 500
   */
  message: string;
};

export type AdminCreateNewsBody = {
  /**
   * @minLength 1
   * @maxLength 100
   */
  title: string;
  /**
   * @minLength 1
   * @maxLength 1000
   */
  body: string;
};

export type AdminSetAdminBody = {
  isAdmin: boolean;
};

export type AdminBanChatBody = {
  /** @minimum 0 */
  minutes: number;
};

export type PlaySlotBody = {
  /** @minimum 1 */
  bet: number;
};

export type PlaySpinBody = {
  /** @minimum 1 */
  bet: number;
};

export type PlayLuckyBoxBody = {
  /** @minimum 1 */
  bet: number;
  /**
   * @minimum 0
   * @maximum 8
   */
  pickIndex: number;
};

export type PlayCrashBody = {
  /** @minimum 1 */
  bet: number;
  /**
   * @minimum 1.01
   * @maximum 50
   */
  autoCashout: number;
};

export type GetChatMessagesParams = {
  /**
   * @maximum 100
   */
  limit?: number;
};

export type PostChatMessageBody = {
  /**
   * @minLength 1
   * @maxLength 200
   */
  message: string;
};

export type AdminLoginBody = {
  password: string;
};

export type AdminListUsersParams = {
  search?: string;
};

export type AdminAdjustCoinsBody = {
  amount: number;
  /**
   * @minLength 1
   * @maxLength 200
   */
  reason: string;
};

export type AdminListTransactionsParams = {
  userId?: string;
};