# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CoinHub API - virtual coin economy platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Täze ulanyjy döretmek
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 24
                password:
                  type: string
                  minLength: 4
                  maxLength: 64
                email:
                  type: string
                  format: email
              required: [username, password]
      responses:
        "200":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Ulanyjy girmek
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required: [username, password]
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /admin/users/{userId}/set-admin:
    post:
      operationId: adminSetAdmin
      tags: [admin]
      summary: Ulanyjyny admin owner edip belgilemek
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isAdmin:
                  type: boolean
              required: [isAdmin]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /admin/users/{userId}/ban-chat:
    post:
      operationId: adminBanChat
      tags: [admin]
      summary: Ulanyjyny çatda gadagan etmek
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                minutes:
                  type: integer
                  minimum: 0
              required: [minutes]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
  /chat/messages:
    get:
      operationId: getChatMessages
      tags: [chat]
      summary: Soňky çat habarlary
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: postChatMessage
      tags: [chat]
      summary: Habar ibermek
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                  minLength: 1
                  maxLength: 200
              required: [message]
      responses:
        "200":
          description: Posted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
  /admin/chat/messages/{messageId}:
    delete:
      operationId: adminDeleteChatMessage
      tags: [admin]
      summary: Çat habary öçürmek
      parameters:
        - in: path
          name: messageId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin girmek
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
              required: [password]
      responses:
        "200":
          description: Admin logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: Ähli ulanyjylar
      parameters:
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
  /admin/users/{userId}:
    get:
      operationId: adminGetUser
      tags: [admin]
      summary: Ulanyjyny görmek
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
  /admin/users/{userId}/coins:
    post:
      operationId: adminAdjustCoins
      tags: [admin]
      summary: Ulanyja teňňe goşmak / aýyrmak
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: integer
                reason:
                  type: string
                  minLength: 1
                  maxLength: 200
              required: [amount, reason]
      responses:
        "200":
          description: Adjusted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
  /admin/transactions:
    get:
      operationId: adminListTransactions
      tags: [admin]
      summary: Ähli amallar taryhy
      parameters:
        - in: query
          name: userId
          schema:
            type: string
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
...
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
      required: [user]
    User:
      type: object
      properties:
        id:
          type: string
        publicId:
          type: string
        username:
          type: string
        coins:
          type: integer
        createdAt:
          type: string
        isAdmin:
          type: boolean
        email:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        avatarColor:
          type: string
        bonusReady:
          type: boolean
        unreadNotifications:
          type: integer
        unreadDms:
          type: integer
      required: [id, publicId, username, coins, createdAt, isAdmin, avatarColor, bonusReady, unreadNotifications, unreadDms]
    Transaction:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        amount:
          type: integer
        reason:
          type: string
        source:
          type: string
        createdAt:
          type: string
      required: [id, userId, amount, reason, source, createdAt]
...
    ChatMessage:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        publicId:
          type: string
        username:
          type: string
        avatarColor:
          type: string
        isAdmin:
          type: boolean
        message:
          type: string
        createdAt:
          type: string
      required: [id, userId, publicId, username, avatarColor, isAdmin, message, createdAt]
...
    AdminUser:
      type: object
      properties:
        id:
          type: string
        publicId:
          type: string
        username:
          type: string
        coins:
          type: integer
        createdAt:
          type: string
      required: [id, publicId, username, coins, createdAt]
    AdminUserDetail:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/AdminUser"
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
      required: [user, transactions]
...