# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CoinHub API - virtual coin economy platform
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  publicId: string;
  username: string;
  coins: number;
  createdAt: string;
  isAdmin: boolean;
  email?: string | null;
  bio?: string | null;
  avatarColor: string;
  bonusReady: boolean;
  unreadNotifications: number;
  unreadDms: number;
}